Update app/Controller/Component/CartComponent.php
I had an issue, where if uses go back to the products page and decided to add another item to the cart and that item was already in the cart, that item would simply get replaced with the new entry. I wanted to actually add to the quantity of this product ID if it is already in the cart. Maybe there is an easier way, but this way works for me and it is the only idea I have at this time.